rename edit page
